Increasing rubycritics score.
